backend/pyproject.toml
[tool.poetry]
name = "beacon-library-backend"
version = "0.1.0"
description = "Backend API for Beacon Library Electronic Document Management System"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"

# Storage (MinIO/S3)
aioboto3 = "^13.0.0"
types-aiobotocore = "^2.12.0"

# Authentication (Keycloak)
python-keycloak = "^4.0.0"
authlib = "^1.3.0"
httpx = "^0.27.0"

# Caching (Redis)
redis = "^5.0.0"

# MCP Server
mcp = "^1.0.0"

# Email
aiosmtplib = "^3.0.0"

# Observability dependencies
prometheus-fastapi-instrumentator = "^7.0.0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-exporter-otlp = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-instrumentation-logging = "^0.48b0"
opentelemetry-instrumentation-sqlalchemy = "^0.48b0"
opentelemetry-instrumentation-redis = "^0.48b0"
structlog = "^24.4.0"

# SSE for real-time updates
sse-starlette = "^2.0.0"

# Semantic search (future epics)
chromadb = "^0.4.0"
ollama = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
